Initial CI/CD Pipeline (#3)
This merge adds an initial CI/CD pipeline file to run tests on PRs and
pushes to `main`, along with a workflow dispatch trigger. Tests execute
on Windows, Ubuntu, and MacOS, with the latest couple versions of Go
(1.22, 1.23).

The regex previously used to match output in end-to-end tests worked for
unix-based systems, but was incorrect for Windows. Rather than
introducing conditional checks to the test, the regex has been
simplified to pass on all platforms.